# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentation",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "variables": {}
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "create user",
        "operationId": "createuser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createuserrequest"
                  },
                  {
                    "example": {
                      "username": "parangaricurimicuaro",
                      "email": "parangaricurimicuaro@gmail.com",
                      "password": "Paranga@23"
                    }
                  }
                ]
              },
              "example": {
                "username": "parangaricurimicuaro",
                "email": "parangaricurimicuaro@gmail.com",
                "password": "Paranga@23"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "login",
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginrequest"
                  },
                  {
                    "example": {
                      "email": "parangaricurimicuaro@gmail.com",
                      "password": "Paranga@23"
                    }
                  }
                ]
              },
              "example": {
                "email": "parangaricurimicuaro@gmail.com",
                "password": "Paranga@23"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "upload avatar",
        "operationId": "uploadavatar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "encoding": {},
              "schema": {
                "required": [
                  "avatar"
                ],
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/products": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get all product",
        "operationId": "Getallproduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "get product by id",
        "operationId": "getproductbyid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "producto id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/products/car/{id}": {
      "post": {
        "tags": [
          "car"
        ],
        "summary": "add product to car",
        "operationId": "addproducttocar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "car id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "id"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/addproducttocarrequest"
                  },
                  {
                    "example": {
                      "productId": 5,
                      "quantity": 2,
                      "price": 19.49
                    }
                  }
                ]
              },
              "example": {
                "productId": 5,
                "quantity": 2,
                "price": 19.49
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "createuserrequest": {
        "title": "createuserrequest",
        "required": [
          "username",
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "username": "parangaricurimicuaro",
          "email": "parangaricurimicuaro@gmail.com",
          "password": "Paranga@23"
        }
      },
      "loginrequest": {
        "title": "loginrequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "parangaricurimicuaro@gmail.com",
          "password": "Paranga@23"
        }
      },
      "addproducttocarrequest": {
        "title": "addproducttocarrequest",
        "required": [
          "productId",
          "quantity",
          "price"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number"
          }
        },
        "example": {
          "productId": 5,
          "quantity": 2,
          "price": 19.49
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
  "tags": [
    {
      "name": "users"
    },
    {
      "name": "products"
    },
    {
      "name": "car"
    }
  ]
}